Remove old UserController
