+++ app/src/main/java/org/hsse/news/api/SparkApplication.java
@@ -6,7 +6,6 @@ import org.hsse.news.api.authorizers.BasicAuthorizer;
 import org.hsse.news.api.controllers.ArticleController;
 import org.hsse.news.api.controllers.Controller;
 import org.hsse.news.api.controllers.TopicController;
-import org.hsse.news.api.controllers.UserController;
 import org.hsse.news.api.controllers.WebsiteController;
 import org.hsse.news.database.article.ArticleService;
 import org.hsse.news.database.topic.TopicService;
@@ -36,7 +35,6 @@ public class SparkApplication implements SubApplication {
         this.controllers = List.of(
                 new ArticleController(API_PREFIX, service, articleService, websiteService, topicService, objectMapper, authorizer),
                 new TopicController(API_PREFIX, service, topicService, objectMapper, authorizer),
-                new UserController(API_PREFIX, service, userService, objectMapper, authorizer),
                 new WebsiteController(API_PREFIX, service, websiteService, objectMapper, authorizer)
         );
     }
